[Build] Move test configurations from 'common-configurations' into 'project-tests-convention'
