+++ repo/gradle-build-conventions/project-tests-convention/src/main/kotlin/TestInventoryListener.kt
