+++ repo/gradle-build-conventions/project-tests-convention/src/main/kotlin/configureTestRetries.kt
@@ -1,14 +1,17 @@
 /*
- * Copyright 2010-2023 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
 import org.gradle.api.Project
 import org.gradle.api.tasks.testing.Test
 import org.gradle.kotlin.dsl.develocity
+import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.withType
 
-fun Project.configureTestRetriesForTestTasks() {
+internal fun Project.configureTestRetries() {
+    val extension = project.extensions.getByType<ProjectTestsExtension>()
+
     val testRetryMaxRetries = kotlinBuildProperties.intProperty("kotlin.build.testRetry.maxRetries").orNull
         ?: (if (kotlinBuildProperties.isTeamcityBuild.get()) 3 else 0)
 
@@ -17,6 +20,7 @@ fun Project.configureTestRetriesForTestTasks() {
             maxRetries.set(testRetryMaxRetries)
             maxFailures.set(20)
             failOnPassedAfterRetry.set(false)
+            failOnPassedAfterRetry.set(extension.allowFlaky.convention(true).map { !it })
         }
     }
 }
