+++ repo/gradle-build-conventions/project-tests-convention/src/main/kotlin/confiugreTestInventory.kt
@@ -1,0 +1,17 @@
+import org.gradle.api.Project
+import org.gradle.api.tasks.testing.AbstractTestTask
+import org.gradle.kotlin.dsl.withType
+
+/*
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+internal fun Project.configureTestInventory() {
+    tasks.withType<AbstractTestTask>().configureEach {
+
+        val testInventoryListener = TestInventoryListener(name, project.layout.buildDirectory.asFile)
+        addTestListener(testInventoryListener)
+        outputs.file(testInventoryListener.inventoryFile)
+    }
+}
