+++ repo/gradle-build-conventions/test-federation-convention/src/main/kotlin/org/jetbrains/kotlin/test/federation/TestFederationExtension.kt
@@ -10,9 +10,5 @@ import org.gradle.api.provider.Property
 import javax.inject.Inject
 
 open class TestFederationExtension @Inject constructor(project: Project) {
-    /**
-     * The module 'repo:test-runtime' is added to test's as a default dependency.
-     * This property can be used to disable this default.
-     */
-    val defaultDependencyEnabled: Property<Boolean> = project.objects.property(Boolean::class.java).convention(true)
+
 }
